[feat] #27 : 로또 6/45 entity class 설계 변경
스크랩핑 기능을 개발하고, 테스트하던 중 db 구조가 변경되어야 할 필요가 있어서 entity class 설계를 변경했다.

+++ src/main/java/com/example/projectlottery/domain/id/LottoWinShopId.java
@@ -9,6 +9,6 @@ import java.io.Serializable;
 public class LottoWinShopId implements Serializable {
 
     private Lotto lotto;
-    private Long rank;
-    private Long no;
+    private Integer rank;
+    private Integer no;
 }
